backend/requirements.txt
openai
fastapi
uvicorn
python-multipart
pydantic
python-dotenv